Move classes to Gyoza project and update references & namespace
