+++ test/Gyoza.Tests/Gyoza.Tests.csproj
@@ -11,6 +11,7 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
     <PackageReference Include="xunit" Version="2.2.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
+    <ProjectReference Include="../../src/Gyoza/Gyoza.csproj" />
   </ItemGroup>
 
 </Project>
